+++ sdk/core/src/main/java/com/klaviyo/core/config/Log.kt
@@ -1,6 +1,6 @@
 package com.klaviyo.core.config
 
-import java.net.URL
+import com.klaviyo.core.networking.NetworkRequest
 
 interface Log {
 
@@ -66,66 +66,4 @@ interface Log {
      * @param value Null if the value is being cleared
      */
     fun onDataStore(key: String, value: String?)
-}
-
-/**
- * Immutable representation of the data of a network request
- */
-interface NetworkRequest {
-    /**
-     * Unsent, PendingRetry, Complete or Failed
-     */
-    val state: String
-
-    /**
-     * Time the request was initiated
-     */
-    val startTime: String
-
-    /**
-     * Time the request was completed or failed
-     */
-    val endTime: String?
-
-    /**
-     * URL of the request, omitting query string
-     */
-    val url: URL
-
-    /**
-     * GET or POST
-     */
-    val httpMethod: String
-
-    /**
-     * HTTP Headers
-     */
-    val headers: Map<String, String>
-
-    /**
-     * Query string represented as dictionary
-     */
-    val query: Map<String, String>
-
-    /**
-     * Render the POST body a string
-     *
-     * @return
-     */
-    fun formatBody(): String?
-
-    /**
-     * Render the response as a string
-     * Format depends on the endpoint, see Klaviyo API documentation
-     *
-     * @return
-     */
-    fun formatResponse(): String?
-
-    /**
-     * Render the whole request object as a JSON string
-     *
-     * @return
-     */
-    override fun toString(): String
 }
