+++ sdk/core/src/main/java/com/klaviyo/core/networking/NetworkRequest.kt
@@ -1,0 +1,65 @@
+package com.klaviyo.core.networking
+
+import java.net.URL
+
+/**
+ * Immutable representation of the data of a network request
+ */
+interface NetworkRequest {
+    /**
+     * Unsent, PendingRetry, Complete or Failed
+     */
+    val state: String
+
+    /**
+     * Time the request was initiated
+     */
+    val startTime: String
+
+    /**
+     * Time the request was completed or failed
+     */
+    val endTime: String?
+
+    /**
+     * URL of the request, omitting query string
+     */
+    val url: URL
+
+    /**
+     * GET or POST
+     */
+    val httpMethod: String
+
+    /**
+     * HTTP Headers
+     */
+    val headers: Map<String, String>
+
+    /**
+     * Query string represented as dictionary
+     */
+    val query: Map<String, String>
+
+    /**
+     * Render the POST body a string
+     *
+     * @return
+     */
+    fun formatBody(): String?
+
+    /**
+     * Render the response as a string
+     * Format depends on the endpoint, see Klaviyo API documentation
+     *
+     * @return
+     */
+    fun formatResponse(): String?
+
+    /**
+     * Render the whole request object as a JSON string
+     *
+     * @return
+     */
+    override fun toString(): String
+}
